requirements.txt.txt
python-telegram-bot==20.3
psycopg2-binary==2.9.9
pandas==2.1.4
requests==2.31.0
python-dotenv==1.0.0
nest-asyncio==1.5.8
sqlalchemy==2.0.25